# vegafixed/vega-spring/src/main/java/io/venus/vega/api/v1/FileController.java
package io.venus.vega.api.v1;

import io.venus.vega.api.v1.resources.FileDownloadRequestResource;
import io.venus.vega.api.v1.resources.FileListResource;
import io.venus.vega.services.FileService;
import lombok.RequiredArgsConstructor;
import org.springframework.core.io.Resource;
import org.springframework.http.HttpHeaders;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.multipart.MultipartFile;

import java.util.List;

@RestController
@RequiredArgsConstructor
public class FileController extends AbstractController implements FilesApi {

    private final FileService fileService;

    @Override
    public ResponseEntity<Void> createFile(final MultipartFile file) {
        this.fileService.createFile(file);
        return ResponseEntity.status(HttpStatus.CREATED)
                .build();
    }

    @Override
    public ResponseEntity<Resource> downloadFile(String  filePath) {
        final var downloadFileDto = this.fileService.downloadFile(filePath);
        return ResponseEntity.ok()
                .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=" + downloadFileDto.getFileName())
                .body(downloadFileDto.getFile());
    }

    @Override
    public ResponseEntity<List<FileListResource>> getFiles(final Integer offset, final Integer limit, final String sort, final String filter) {
        final var filesPage = this.fileService.getFiles(specificationFrom(filter), pageOf(offset, limit, sortBy(sort)));
        return ResponseEntity.ok()
                .header(TOTAL_COUNT_HEADER, String.valueOf(filesPage.getTotalElements()))
                .header(TOTAL_PAGES_COUNT_HEADER, String.valueOf(filesPage.getTotalPages()))
                .body(filesPage.getContent());
    }
}